Backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
flask-talisman==1.1.0
marshmallow==3.26.1
python-dotenv==1.0.1
pandas==2.3.3
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
networkx==3.4.2
requests==2.33.1
joblib==1.5.3
google-generativeai==0.8.4
google-cloud-logging==3.11.4
google-cloud-translate==3.16.0
flask-limiter==3.7.0
gunicorn==23.0.0
